update : correctAnswer 필드 자료형 변경 및 update 메소드 추가

+++ src/main/java/com/dsm/newtrash/back/springboot/domain/problem/presentation/dto/response/ProblemResponses.java
@@ -19,9 +19,10 @@ public class ProblemResponses {
 		private final Long id;
 		private final String form;
 		private final String question;
-		private final int correctAnswer;
+		private final long correctAnswer;
 		private final String image;
 		private final List<AnswerResponse> answers;
+
 	}
 
 }
